requirements.txt
python-telegram-bot==21.0.1
telethon==1.35.0
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
aiohttp>=3.9.0
httpx>=0.27.0
